feat: start\stop race, create added winner to DB; fix: pagination to create new car
