+++ src/components/app/app.scss
@@ -1,80 +1,76 @@
 *,
 *::after,
 *::before {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
 }
 
 img,
 svg {
-    width: 100%;
-    height: 100%;
+  width: 100%;
+  height: 100%;
 }
 
 body {
-    padding: 0 10px;
+  padding: 0 10px;
 }
 
 button {
-    padding: 5px 10px;
-    border: 1px solid rgb(193, 166, 166);
-    border-radius: 5px;
+  padding: 5px 10px;
+  border: 1px solid rgb(193, 166, 166);
+  border-radius: 5px;
 }
 
 input {
-    border-radius: 5px;
-    border: 1px solid #000;
+  border-radius: 5px;
+  border: 1px solid #000;
 }
 
-// .nav {
-// }
-
 .fields {
-    margin: 20px 0;
+  margin: 20px 0;
 }
 
 .garage {
-    h1 {
-        margin: 10px 0;
-    }
-    .page {
-        margin: 10px 0;
-    }
+  h1 {
+    margin: 10px 0;
+  }
+  .page {
+    margin: 10px 0;
+  }
 }
 
 .car-track {
-    border-bottom: 2px dashed #000;
-    position: relative;
+  border-bottom: 2px dashed #000;
+  position: relative;
 }
 
 .car {
-    width: 100px;
-    height: 50px;
-    margin-bottom: -8px;
-    // transition: 3s;
-    transform: translateX(calc(-100px + 100%));
+  width: 100px;
+  height: 50px;
+  margin-bottom: -8px;
+  transform: translateX(calc(-100px + 100%));
 }
 
 .finish {
-    right: 100px;
-    bottom: -10px;
-    position: absolute;
-    width: 50px;
-    height: 50px;
+  right: 100px;
+  bottom: -10px;
+  position: absolute;
+  width: 50px;
+  height: 50px;
 }
 
 @keyframes drive {
-    from {
-        transform: translateX(calc(-100px + 100%));
-    }
-    to {
-        transform: translateX(calc(-100px + 98vw));
-    }
+  from {
+    transform: translateX(calc(-100px + 100%));
+  }
+  to {
+    transform: translateX(calc(-100px + 98vw));
+  }
 }
+
 .car.drive {
-    animation: drive;
-    // animation-duration: 13s;
-    animation-timing-function: linear;
-    animation-fill-mode: forwards;
+  animation: drive;
+  animation-timing-function: linear;
+  animation-fill-mode: forwards;
 }
