+++ src/components/app/app.ts
@@ -7,9 +7,11 @@ export class App {
     this.garagePageCounter = 1;
     this.tracksOnPageAmount = 7;
   }
+
   get garagePage() {
     return this.garagePageCounter;
   }
+
   set garagePage(pageNumber: number) {
     if (pageNumber < 1) {
       alert('page cant be less 1');
@@ -17,6 +19,7 @@ export class App {
     }
     this.garagePageCounter = pageNumber;
   }
+
   get tracksOnPage() {
     return this.tracksOnPageAmount;
   }
