+++ src/components/race/race.ts
@@ -1,6 +1,7 @@
 import { ICar } from '../../types';
 import { API } from '../api/api';
-import { CarTrack } from '../car-track/carTrack';
+import { App } from '../app/app';
+import { CarTrack } from '../carTrack/carTrack';
 
 export class Race {
   generateButtons() {
@@ -13,6 +14,7 @@ export class Race {
         `;
     return block;
   }
+
   generateCars(carTrack: CarTrack, count = 100) {
     const carNames = [
       'Lada',
@@ -41,15 +43,16 @@ export class Race {
       carTrack.createCar(carNames[car], `#${color}`);
     }
   }
-  generateRaceListeners(carTrack: CarTrack, api: API) {
+
+  generateRaceListeners(carTrack: CarTrack, api: API, app: App) {
     const generateCarsButton = document.querySelector('.fields__button-generate');
+
     generateCarsButton?.addEventListener('click', () => {
       this.generateCars(carTrack);
-      carTrack.createTrack(api.getCars<ICar[]>('http://127.0.0.1:3000/garage', 1));
+      carTrack.createTrack(api.getCars(1));
       carTrack.updateGarageAmount(api);
-      setTimeout(()=>{
-        carTrack.carHandler(api);
-      }, 100)
+      setTimeout(() => carTrack.carHandler(api), 300);
+      carTrack.paginationClickableButtons(app, api);
     });
   }
 }
